requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
asyncpg==0.29.0
alembic==1.13.2
pydantic-settings==2.4.0
python-dotenv==0.21.0
psycopg2-binary==2.9.7